doc: Remove Sun references and URLs
The docs still contain references to Sun and Sun URLs. They should be
updated to reference Oracle and Oracle URLs. This PR doesn't fix all of
the Sun references and URLs. The "escapes" chapters still contains URLs
and references to the "JDBC Technology Guide" and
"JDBC 3.0 specification" which I don't know what to replace with.